hash for 'x' conversion bug fixed

+++ lib/ft_itoa.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-char			*ft_itoa(int n)
+char			*ft_itoa(long long int n)
 {
 	return (itoa_base(n, 10));
 }
